% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp.R
\name{directionalPalette}
\alias{directionalPalette}
\title{Directional color palette.}
\usage{
directionalPalette(n = 64, col = "green3", p = 0.8, x = NULL, ...)
}
\arguments{
\item{n}{Number of colors}

\item{col}{Base color for the palette.}

\item{p}{Power to scale the increase in intensity. Default is 0.8, 1 gives linear decrease.}
}
\value{

}
\description{
Directional color palette.
}
