% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_selectivity.R
\name{fit_selectivity}
\alias{fit_selectivity}
\title{Title}
\usage{
fit_selectivity(
  object,
  pattern = 0,
  blocks = NULL,
  w = NULL,
  method = NULL,
  control = list(),
  FUN = NULL,
  ...
)
}
\arguments{
\item{object}{}

\item{...}{}
}
\description{
Title
}
