% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_selectivity.R
\name{predict.selectivity_model}
\alias{predict.selectivity_model}
\alias{predict.empirical_selectivity_model}
\title{Predict the selectivity for a new length or age vector.}
\usage{
\method{predict}{selectivity_model}(object, x, ...)

\method{predict}{empirical_selectivity_model}(object, x, ...)
}
\arguments{
\item{object}{A 'selectivity_model' or 'empirical_selectivity_model' object.}

\item{x}{The length or age vector.}

\item{...}{Additional arguments, currently not used.}
}
\description{
Predict the selectivity for a new length or age vector.
}
