% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_selectivity-methods.R
\name{split.empirical_selectivity}
\alias{split.empirical_selectivity}
\title{Title}
\usage{
\method{split}{empirical_selectivity}(x, f, breaks, drop = FALSE, sep = ".", lex.order = FALSE, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
Title
}
