% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_selectivity-methods.R
\name{[.empirical_selectivity}
\alias{[.empirical_selectivity}
\title{Title}
\usage{
\method{[}{empirical_selectivity}(x, i, j, yr, age, length, drop = FALSE)
}
\arguments{
\item{x}{}

\item{drop}{}
}
\description{
Title
}
