% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_selectivity-methods.R
\name{weighted.mean.empirical_selectivity}
\alias{weighted.mean.empirical_selectivity}
\title{Title}
\usage{
\method{weighted.mean}{empirical_selectivity}(x, w, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
Title
}
