% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_selectivity.R
\name{predict.selectivity_model}
\alias{predict.selectivity_model}
\title{Title}
\usage{
\method{predict}{selectivity_model}(object, x, ...)
}
\arguments{
\item{object}{}

\item{...}{}
}
\description{
Title
}
